+++ lib/rubocop/cop/flexport/global_model_access_from_engine.rb
@@ -45,8 +45,14 @@ module RuboCop
       #     # No direct association to global models.
       #   end
       #
+      # This cop will also complain if you try to use global FactoryBot
+      # factories in your engine's specs. To disable this behavior for your
+      # engine, add it to the `AllowGlobalFactoryBotFromEngines` list in
+      # .rubocop.yml.
+      #
       class GlobalModelAccessFromEngine < Cop
         include EngineNodeContext
+        include FactoryBotUsage
 
         MSG = 'Direct access of global model `%<model>s` ' \
               'from within Rails Engine.'
@@ -54,6 +60,10 @@ module RuboCop
         def_node_matcher :rails_association_hash_args, <<-PATTERN
           (send _ {:belongs_to :has_one :has_many} sym $hash)
         PATTERN
+
+        class << self
+          attr_accessor :global_factories_cache
+        end
 
         def on_const(node)
           return unless in_enforced_engine_file?
@@ -69,19 +79,41 @@ module RuboCop
           return unless in_enforced_engine_file?
 
           rails_association_hash_args(node) do |assocation_hash_args|
-            class_name_node = extract_class_name_node(assocation_hash_args)
-            class_name = class_name_node&.value
-            next unless global_model?(class_name)
-
-            add_offense(class_name_node, message: message(class_name))
+            check_for_rails_association_with_global_model(assocation_hash_args)
           end
+
+          return unless check_for_global_factory_bot?
+
+          factory_bot_usage(node) do |factory_node|
+            check_for_global_factory_bot_usage(node, factory_node)
+          end
+        end
+
+        def check_for_rails_association_with_global_model(assocation_hash_args)
+          class_name_node = extract_class_name_node(assocation_hash_args)
+          class_name = class_name_node&.value
+          return unless global_model?(class_name)
+
+          add_offense(class_name_node, message: message(class_name))
+        end
+
+        def check_for_global_factory_bot_usage(node, factory_node)
+          factory = factory_node.children[0]
+          return unless global_factory?(factory)
+
+          model_class_name = global_factories[factory]
+          add_offense(node, message: message(model_class_name))
         end
 
         # Because this cop's behavior depends on the state of external files,
         # we override this method to bust the RuboCop cache when those files
         # change.
         def external_dependency_checksum
-          Digest::SHA1.hexdigest(model_dir_paths.join)
+          if check_for_global_factory_bot?
+            Digest::SHA1.hexdigest((model_dir_paths + global_factories.keys.sort).join)
+          else
+            Digest::SHA1.hexdigest(model_dir_paths.join)
+          end
         end
 
         private
@@ -94,8 +126,24 @@ module RuboCop
           @global_model_names ||= calculate_global_models
         end
 
+        def global_factories
+          # Cache factories at the class level so that we don't have to fetch
+          # them again for every file we lint.
+          self.class.global_factories_cache ||= spec_factory_paths.each_with_object({}) do |path, h|
+            source_code = File.read(path)
+            source = RuboCop::ProcessedSource.new(source_code, RUBY_VERSION.to_f)
+            find_factories(source.ast).each do |factory, model_class_name|
+              h[factory] = model_class_name
+            end
+          end
+        end
+
         def model_dir_paths
           Dir[File.join(global_models_path, '**/*.rb')]
+        end
+
+        def spec_factory_paths
+          @spec_factory_paths ||= Dir['spec/factories/**/*.rb']
         end
 
         def calculate_global_models
@@ -133,6 +181,12 @@ module RuboCop
           end
         end
 
+        def check_for_global_factory_bot?
+          spec_file? && allow_global_factory_bot_from_engines.none? do |engine|
+            processed_source.path.include?(File.join(engines_path, engine, ''))
+          end
+        end
+
         def global_model_const?(const_node)
           # Remove leading `::`, if any.
           class_name = const_node.source.sub(/^:*/, '')
@@ -142,6 +196,10 @@ module RuboCop
         # class_name is e.g. "FooGlobalModelNamespace::BarModel"
         def global_model?(class_name)
           global_model_names.include?(class_name)
+        end
+
+        def global_factory?(factory_name)
+          global_factories.include?(factory_name)
         end
 
         def child_of_const?(node)
@@ -163,6 +221,10 @@ module RuboCop
           raw.map do |e|
             ActiveSupport::Inflector.underscore(e)
           end
+        end
+
+        def allow_global_factory_bot_from_engines
+          cop_config['AllowGlobalFactoryBotFromEngines'] || []
         end
 
         def allowed_global_models