+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master (unreleased)
 
+## 0.10.0
+
+- EngineApiBoundary cop detects cross-engine use of FactoryBoy factories.
+- GlobalModelAccessFromEngine cop detects use of global FactoryBoy factories.
+
 ## 0.9.0
 
 - Add support for `_allowlist.rb`, in addition to `_whitelist.rb`.